requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.3.1
pynetworktables==2019.0.2
roslibpy==0.6.0
six==1.12.0
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.4.0
wrapt==1.11.2
zope.interface==4.6.0
